libocheck: split alloc & files messages into their own message stores
This is a big change.
Seems that keeping a 512k store of messages, and iterating through
it (on each call) causes a big performance penalty.

So, then the solution is keep an 'upper_index_limit', which
is more the biggest 'index' which the last limit until where we
had to store messages.
So (for example) if the max number of allocs was 200, we should
not have to go to the end of the array to search for a duplicate
entry.

The problem with duplicate entries, is due to the fact that
realloc() (and possibly calloc() too) calls malloc().
This is pretty dependant on the libc implementation (this may
not be true for musl-libc, but is true for uClibc).

Another optimization done here, is to split the stores for allocs
and files to their own.
Technically, it's unlikely that we'll get more than 64k FDs.
And maybe we won't get more thatn 128k allocs (per program).

But mixing them together into a single store, complicates things,
and it's hard to tell if a limit is reached by allocs or files
(when both are stored in the same place).

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@riverbed.com>