libocheck: disable flushing by default during execution
So, I don't know [anymore] if it makes sense to flush
information during execution [of a program] or just at the
end of the execution.

For sure [during a test] we're interested only at the end of the
execution.

Problem is, the current logic can block the execution of a program
especially when the flush_counter expires.

To fix the blocking issue, we'd need to store info in another
buffer and write that (to ocheckd), while we continue storing
allocs & fds in the main buffer.

But that feature is future development.

For, now we'll make the default to not flush info during
program execution.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@riverbed.com>